% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_functions.R
\name{bootPCA}
\alias{bootPCA}
\title{Quickly calculates bootstrap PCA results (wrapper for bootSVD)}
\usage{
bootPCA(...)
}
\arguments{
\item{...}{passed to \code{\link{bootSVD}}, with centerSamples set to TRUE.}
}
\value{
\code{bootSVD(...)}
}
\description{
All arguments are passed to \code{\link{bootSVD}}. This function should be used in exactly the same way as \code{\link{bootSVD}}. The only difference is that PCA typically involves re-centering each bootstrap sample, whereas calculations involving the SVD might not.
}
